Use randomness for ghost-mode movement.
Okay, so.  The original idea was to make the ghosts always flee; that
worked too well, because they'd flying across the board and be in a
strategically bad position (if I can even call it that).

The next idea was to rotate the direction (move sideways from the
hypothetical pursuer), but that also seemed to result in worse
performance (not that I have anything remotely resembling a statistically
valid sample, but whatever).

And then that didn't work well, so it was sort of random but not really;
but now it is.